+++ modules/cmd_args.py
@@ -113,4 +113,3 @@ parser.add_argument("--logging-level", type=str,  help="logging level, can be on
 parser.add_argument("--model-cache-dir", type=str,  help="User can use a ssd to cache the models to speed up loading speed", default="")
 parser.add_argument("--model-cache-max-size", type=int,  help="The maximum disk space (GB) to use to cache the model. Need to set --model-cache-dir first.", default=0)
 parser.add_argument("--predict-timeout", type=int,  help="Timeout in second of /predict.", default=1800)
-parser.add_argument("--lora-dir", type=str, help="Path to directory with Lora networks.", default=os.path.join(models_path, 'Lora'))
