+++ webui.py
@@ -25,7 +25,6 @@ logging.getLogger("xformers").addFilter(lambda record: 'A matching Triton is not
 from modules import paths, timer, import_hook, errors
 from modules.paths_internal import data_path
 from modules.state_holder import make_state_holder
-from modules.Lora.scripts import lora_script
 
 startup_timer = timer.Timer('main', 'initialize')
 
@@ -219,7 +218,6 @@ def initialize():
     startup_timer.record("list builtin upscalers")
 
     modules.scripts.load_scripts()
-    lora_script.init()
     startup_timer.record("load scripts")
 
     modules.sd_vae.refresh_vae_list()
